fix: Android build errors — clean compile
- Add .gitignore for Android (excludes .gradle, local.properties, build)
- Fix java.util.Properties import in build.gradle.kts
- Fix duplicate release buildType block
- Fix Theme.Material.NoTitleBar → Theme.Material.Light.NoActionBar
- Fix invalid <rect> in vector drawable → <path>
- Fix Float.pow(Double) type mismatch in LocationService
- Fix Map<SocialPlatform, String> type inference in EditProfileScreen
- Fix icon() extension import in OwnProfileScreen
- Add Gradle wrapper (gradlew)

BUILD SUCCESSFUL — 39 Kotlin files, 0 errors

+++ android/app/src/main/java/com/colageclub/colage/core/location/LocationService.kt
@@ -114,7 +114,7 @@ class LocationService @Inject constructor(
         // Convert pressure to altitude using standard barometric formula
         // altitude (m) = 44330 * (1 - (P/P0)^(1/5.255)), P0 = 1013.25 hPa
         val seaLevel = SensorManager.PRESSURE_STANDARD_ATMOSPHERE // 1013.25 hPa
-        val altitudeM = 44330.0 * (1.0 - (pressureHPa / seaLevel).pow(1.0 / 5.255))
+        val altitudeM = 44330.0 * (1.0 - (pressureHPa.toDouble() / seaLevel.toDouble()).pow(1.0 / 5.255))
 
         if (groundAltitude == null) {
             groundAltitude = altitudeM
